refactor post method of subscription process
