+++ src/main/controllers/CaseReferenceNumberSearchResultController.ts
@@ -2,8 +2,10 @@ import { Response } from 'express';
 import { PipRequest } from '../models/request/PipRequest';
 import { cloneDeep } from 'lodash';
 import { PublicationService } from '../service/PublicationService';
+import {SubscriptionService} from "../service/SubscriptionService";
 
 const publicationService = new PublicationService();
+const subscriptionService = new SubscriptionService();
 
 export default class CaseReferenceNumberSearchResultController {
     public async get(req: PipRequest, res: Response): Promise<void> {
@@ -34,5 +36,10 @@ export default class CaseReferenceNumberSearchResultController {
         } else {
             res.render('error', req.i18n.getDataByLanguage(req.lng).error);
         }
+    }
+
+    public async post(req: PipRequest, res: Response): Promise<void> {
+        await subscriptionService.handleNewSubscription(req.body, req.user);
+        res.redirect('/pending-subscriptions');
     }
 }
