+++ src/main/resources/locales/en/subscription-configure-list-language.json
