+++ src/main/routes/routes.ts
@@ -225,6 +225,11 @@ export default function (app: Application): void {
         isPermittedMedia,
         app.locals.container.cradle.caseNameSearchResultsController.get
     );
+    app.post(
+        '/case-name-search-results',
+        isPermittedMedia,
+        app.locals.container.cradle.caseNameSearchResultsController.post
+    );
     app.get(
         '/case-reference-number-search',
         isPermittedMedia,
@@ -239,6 +244,11 @@ export default function (app: Application): void {
         '/case-reference-number-search-results',
         isPermittedMedia,
         app.locals.container.cradle.caseReferenceNumberSearchResultController.get
+    );
+    app.post(
+        '/case-reference-number-search-results',
+        isPermittedMedia,
+        app.locals.container.cradle.caseReferenceNumberSearchResultController.post
     );
     app.get('/delete-subscription', isPermittedMedia, app.locals.container.cradle.deleteSubscriptionController.get);
     app.get(
@@ -258,6 +268,10 @@ export default function (app: Application): void {
     app.post('/location-name-search', isPermittedMedia, (req, res) =>
         app.locals.container.cradle.alphabeticalSearchController.post(req, res, 'location-name-search')
     );
+    app.post(
+        '/location-subscriptions-confirmation',
+        isPermittedMedia,
+        app.locals.container.cradle.alphabeticalSearchController.locationSubscriptionsConfirmation);
     app.get('/pending-subscriptions', isPermittedMedia, app.locals.container.cradle.pendingSubscriptionsController.get);
     app.post(
         '/pending-subscriptions',
@@ -271,12 +285,6 @@ export default function (app: Application): void {
     );
     app.get('/subscription-add', isPermittedMedia, app.locals.container.cradle.subscriptionAddController.get);
     app.post('/subscription-add', isPermittedMedia, app.locals.container.cradle.subscriptionAddController.post);
-    app.get('/subscription-confirmed', isPermittedMedia, reRenderView);
-    app.post(
-        '/subscription-confirmed',
-        isPermittedMedia,
-        app.locals.container.cradle.subscriptionConfirmedController.post
-    );
     app.get(
         '/subscription-management',
         isPermittedMedia,
@@ -293,6 +301,14 @@ export default function (app: Application): void {
         isPermittedMedia,
         app.locals.container.cradle.subscriptionAddListLanguageController.get
     );
+    app.post(
+        '/subscription-add-list-language',
+        isPermittedMedia,
+        app.locals.container.cradle.subscriptionAddListLanguageController.post
+    );
+    app.get('/subscription-confirmed',
+        isPermittedMedia,
+        reRenderView);
     app.get(
         '/subscription-configure-list',
         isPermittedMedia,
@@ -304,16 +320,16 @@ export default function (app: Application): void {
         app.locals.container.cradle.subscriptionConfigureListController.post
     );
     app.get(
-        '/subscription-config-list-language',
+        '/subscription-configure-list-language',
         isPermittedMedia,
-        app.locals.container.cradle.subscriptionConfigListLanguageController.get
+        app.locals.container.cradle.subscriptionConfigureListLanguageController.get
+    );
+    app.post(
+        '/subscription-configure-list-language',
+        isPermittedMedia,
+        app.locals.container.cradle.subscriptionConfigureListLanguageController.post
     );
     app.get('/subscription-configure-list-confirmed', isPermittedMedia, reRenderView);
-    app.post(
-        '/subscription-configure-list-confirmed',
-        isPermittedMedia,
-        app.locals.container.cradle.subscriptionConfigureListConfirmedController.post
-    );
     app.get('/unsubscribe-confirmation', isPermittedMedia, reRenderView);
     app.post(
         '/unsubscribe-confirmation',
