+++ src/main/views/case-name-search-results.njk
@@ -14,7 +14,7 @@
   <div class="parent-box alphabetical-box">
     <h1 class="govuk-heading-l">{{ title }}</h1>
     <p class="govuk-body govuk-!-font-weight-bold">{{ searchResults | length }} {{ resultsCount }}</p>
-    <form action="pending-subscriptions" method="post" autocomplete="off">
+    <form action="case-name-search-results" method="post" autocomplete="off">
         <table class="govuk-table govuk-!-width-three-quarters" data-module="moj-multi-select">
             <thead class="govuk-table__head">
             <tr class="govuk-table__row">
