+++ src/main/views/case-reference-number-search-results.njk
@@ -34,7 +34,7 @@
     </tbody>
   </table>
 
-  <form action="pending-subscriptions" method="post">
+  <form action="case-reference-number-search-results" method="post">
     <div class="govuk-form-group govuk-!-width-one-half">
       {% if urnSearch %}
         {{ govukInput({
