+++ src/main/views/location-name-search.njk
@@ -88,7 +88,7 @@
       </form>
     </div>
     <div class="layout-width-three-fifths" id="content">
-      <form class="moj-action-bar" action="pending-subscriptions" method="post" autocomplete="off">
+      <form class="moj-action-bar" action="location-subscriptions-confirmation" method="post" autocomplete="off">
         <h2 class="govuk-heading-m">{{ userSelections }}</h2>
         <div class="moj-button-menu">
           <div class="moj-button-menu__wrapper">
