+++ src/main/views/pending-subscriptions.njk
@@ -24,7 +24,7 @@
     {% set buttonText = buttonText_court_subscription %}
     {% set formAction = 'get' %}
   {% else %}
-    {% set nextPage = 'subscription-confirmed' %}
+    {% set nextPage = 'pending-subscriptions' %}
     {% set pageTitle = title %}
     {% set buttonText = buttonText %}
     {% set formAction = 'post' %}
