+++ src/main/views/subscription-add-list-language.njk
@@ -29,7 +29,7 @@
             ]
         }) }}
     {% endif %}
-    <form action="subscription-confirmed" method="post">
+    <form action="subscription-add-list-language" method="post">
         <div class="govuk-form-group">
             <fieldset class="govuk-fieldset" id="language-selector">
                 <legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
