+++ src/main/views/subscription-add-list.njk
@@ -57,10 +57,7 @@
                                                 <input type="checkbox" class="govuk-checkboxes__input"
                                                        id="{{ listName }}"
                                                        value="{{ listName}}" name="list-selections[]"
-                                                    {% if listConfig.checked %}
-                                                        checked
-                                                    {% endif %}
-                                                    onclick="calculateChecked()">
+                                                       onclick="calculateChecked()">
                                                 <label class="govuk-label govuk-checkboxes__label no_padding"
                                                        for="{{ listName }}">
                                                     <span class="govuk-visually-hidden">Select {{ listConfig.listFriendlyName }}</span>
