+++ src/main/views/subscription-configure-list-language.njk
@@ -29,7 +29,7 @@
                 ]
             }) }}
         {% endif %}
-        <form action="subscription-configure-list-confirmed" method="post">
+        <form action="subscription-configure-list-language" method="post">
             <div class="govuk-form-group">
                 <fieldset class="govuk-fieldset" id="language-selector">
                     <legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
