+++ src/test/routes/case-reference-number-search-results.test.ts
@@ -13,4 +13,14 @@ describe('Case reference number search result', () => {
                 .expect(res => expect(res.status).to.equal(200));
         });
     });
+
+    describe('on POST', () => {
+        test('should redirect to pending subscription page', async () => {
+            await request(app)
+                .post('/pending-subscriptions')
+                .expect(res => {
+                    expect(res.status).to.equal(302);
+                });
+        });
+    });
 });
