+++ src/test/unit/controllers/SubscriptionConfigureListLanguageController.test.ts
@@ -1,14 +1,18 @@
 import sinon from 'sinon';
 import { Response } from 'express';
 import { mockRequest } from '../mocks/mockRequest';
-import SubscriptionAddListLanguageController from '../../../main/controllers/SubscriptionAddListLanguageController';
-import {SubscriptionService} from "../../../main/service/SubscriptionService";
-import {PendingSubscriptionsFromCache} from "../../../main/service/PendingSubscriptionsFromCache";
+import SubscriptionConfigureListLanguageController from '../../../main/controllers/SubscriptionConfigureListLanguageController';
+import { SubscriptionService } from '../../../main/service/SubscriptionService';
+import { PendingSubscriptionsFromCache } from '../../../main/service/PendingSubscriptionsFromCache';
 
 const userId = '1';
 const language = 'en';
 
-const subscriptionAddListLanguageController = new SubscriptionAddListLanguageController();
+const subscriptionConfigureListLanguageController = new SubscriptionConfigureListLanguageController();
+
+const getSubscriptionListLanguage = sinon.stub(SubscriptionService.prototype, 'getUserSubscriptionListLanguage');
+getSubscriptionListLanguage.withArgs(userId).resolves(['ENGLISH']);
+
 const subscribeStub = sinon.stub(SubscriptionService.prototype, 'subscribe');
 const cacheStub = sinon.stub(PendingSubscriptionsFromCache.prototype, 'getPendingSubscriptions');
 subscribeStub.withArgs('1').resolves(true);
@@ -25,10 +29,9 @@ cacheStub.withArgs('3', 'courts').resolves([]);
 cacheStub.withArgs('3', 'listTypes').resolves([]);
 
 const i18n = {
-    'subscription-add-list-language': {},
+    'subscription-configure-list-language': {},
     error: {},
-}
-;
+};
 const response = {
     render: () => {
         return '';
@@ -38,52 +41,54 @@ const response = {
     },
 } as unknown as Response;
 
-describe('Add List Language Subscriptions Controller', () => {
+describe('Configure List Language Subscriptions Controller', () => {
     describe('GET view', () => {
-        it('should render the Add List Language subscription page without error', () => {
+        it('should render the Configure List Language subscription page without error', () => {
             const request = mockRequest(i18n);
             request.user = { userId: userId };
+            request.query = { error: 'false' };
             request.lng = language;
 
             const expectedData = {
-                ...i18n['subscription-add-list-language'],
+                ...i18n['subscription-configure-list-language'],
+                listTypeLanguageSubscriptions: ['ENGLISH'],
+                noSelectionError: false,
             };
             const responseMock = sinon.mock(response);
-            responseMock.expects('render').once().withArgs('subscription-add-list-language', expectedData);
+            responseMock.expects('render').once().withArgs('subscription-configure-list-language', expectedData);
 
-            return subscriptionAddListLanguageController.get(request, response).then(() => {
+            return subscriptionConfigureListLanguageController.get(request, response).then(() => {
                 responseMock.verify();
             });
         });
     });
-
     describe('POST view', () => {
-        it('should redirect subscription confirmation page if post data is provided', () => {
+        it('should render confirmed page if subscribed successfully', () => {
             const request = mockRequest(i18n);
-            request.user = { userId: userId };
-            request.lng = language;
+            request.user = { userId: '1' };
             request.body = { 'list-language': 'test' };
 
             const responseMock = sinon.mock(response);
-            responseMock.expects('redirect').once().withArgs('/subscription-confirmed');
+            responseMock.expects('redirect').once().withArgs('/subscription-configure-list-confirmed');
 
-            return subscriptionAddListLanguageController.post(request, response).then(() => {
+            subscriptionConfigureListLanguageController.post(request, response).then(() => {
                 responseMock.verify();
             });
         });
 
-        it('should render subscription Add List Language with error if post data is provided and no selection was select', () => {
+        it('should render subscription Config List Language with error if post data is provided and no selection was select', () => {
             const request = mockRequest(i18n);
             request.user = { userId: userId };
             request.lng = language;
             const expectedData = {
-                ...i18n['subscription-add-list-language'],
+                ...i18n['subscription-configure-list-language'],
+                listTypeLanguageSubscriptions: ['ENGLISH'],
                 noSelectionError: true,
             };
             const responseMock = sinon.mock(response);
-            responseMock.expects('render').once().withArgs('subscription-add-list-language', expectedData);
+            responseMock.expects('render').once().withArgs('subscription-configure-list-language', expectedData);
 
-            return subscriptionAddListLanguageController.post(request, response).then(() => {
+            return subscriptionConfigureListLanguageController.post(request, response).then(() => {
                 responseMock.verify();
             });
         });
@@ -98,28 +103,7 @@ describe('Add List Language Subscriptions Controller', () => {
                 .expects('render')
                 .once()
                 .withArgs('error', { ...i18n.error });
-            subscriptionAddListLanguageController.post(request, response).then(() => {
-                responseMock.verify();
-            });
-        });
-
-        it('should redirect to pending subscriptions if there are no cached subscriptions', () => {
-            const request = mockRequest(i18n);
-            request.user = { userId: '3' };
-            request.body = { 'list-language': 'test' };
-            const response = {
-                render: () => {
-                    return '';
-                },
-                redirect: () => {
-                    return '';
-                },
-            } as unknown as Response;
-            const responseMock = sinon.mock(response);
-
-            responseMock.expects('redirect').once().withArgs('pending-subscriptions?no-subscriptions=true');
-
-            subscriptionAddListLanguageController.post(request, response).then(() => {
+            subscriptionConfigureListLanguageController.post(request, response).then(() => {
                 responseMock.verify();
             });
         });
